Extract trellis engine into exp/pecos-trellis; decouple pecos-bp-trellis from the port
