+++ Cargo.toml
@@ -204,6 +204,7 @@ pecos-experimental = { version = "0.2.0-dev.0", path = "exp/pecos-experimental" 
 pecos-foreign = { version = "0.2.0-dev.0", path = "crates/pecos-foreign" }
 pecos-bp-trellis = { version = "0.2.0-dev.0", path = "exp/pecos-bp-trellis" }
 pecos-frontier = { version = "0.2.0-dev.0", path = "exp/pecos-frontier" }
+pecos-trellis = { version = "0.2.0-dev.0", path = "exp/pecos-trellis" }
 pecos-fusion-blossom = { version = "0.2.0-dev.0", path = "crates/pecos-fusion-blossom" }
 pecos-gpu-sims = { version = "0.2.0-dev.0", path = "crates/pecos-gpu-sims" }
 pecos-hugr = { version = "0.2.0-dev.0", path = "crates/pecos-hugr" }
