+++ exp/pecos-frontier/tests/MUTANTS.md
@@ -4,6 +4,9 @@ The repository has no mutation runner. Apply each compiling edit separately,
 run the named killer test(s), and restore the source before trying the next
 row. “Equivalent” rows document mutations that intentionally have no killer.
 
+Engine-edit rows target `exp/pecos-trellis/src/lib.rs`; the committee-ranking
+row remains in `exp/pecos-frontier/src/lib.rs`.
+
 | Mutant | Exact edit (quoted old → new) | Expected killer test(s) or disposition |
 |---|---|---|
 | `merge_to_max` | `".and_modify(|mass| *mass = logaddexp(*mass, log_mass))"` → `".and_modify(|mass| *mass = (*mass).max(log_mass))"` | `degeneracy_mass_beats_the_single_most_likely_error`; `decode_outputs_match_bitwise_snapshot` |
@@ -12,17 +15,17 @@ row. “Equivalent” rows document mutations that intentionally have no killer.
 | `k_off_by_one` | `"let within_k = index < k;"` → `"let within_k = index <= k;"` | `width_pruning_accounts_for_the_discarded_state_and_mass`; `width_and_delta_pruning_can_change_the_logical_answer`; `decode_outputs_match_bitwise_snapshot` |
 | `remove_forced_seed` | `"let mut initial_syndrome = self.forced_syndrome.clone();"` → `"let mut initial_syndrome = vec![0; self.detector_words];"` | `forced_syndrome_shifts_shared_probabilistic_detector`; `unpruned_matches_independent_brute_force_on_seeded_random_dems` |
 | `swap_suffix_rho_zero_and_one` | In `suffix_compatibility_score`, replace `"row.log_probability_zero"` with `"row.log_probability_one"` and the existing `"row.log_probability_one"` with `"row.log_probability_zero"` in the opposite branch. | `suffix_compatibility_changes_the_greedy_survivor`; `decode_outputs_match_bitwise_snapshot` |
-| `eta_off_by_one_apply_moment_before_snapshot` | Move the complete `"for detector in set_bits(&column.detector_toggle) { row_moments[detector] *= moment; }"` loop from after the `column.suffix_compatibility = ...collect();` assignment to immediately before it. | `suffix_compatibility_changes_the_greedy_survivor`; `decode_outputs_match_bitwise_snapshot` |
+| `eta_off_by_one_apply_moment_before_snapshot` | Move the complete `"for detector in set_bits(&column.detector_toggle) { row_moments[detector] *= moment; }"` loop from after the assignment beginning `"*table = set_bits(&column.active_mask)"` and ending `".collect();"` to immediately before that assignment. | `suffix_compatibility_changes_the_greedy_survivor`; `decode_outputs_match_bitwise_snapshot` |
 | `score_fold_restore` | `"retained.insert(scored.candidate.key, scored.candidate.log_mass);"` → `"retained.insert(scored.candidate.key, scored.score);"` | `decode_outputs_match_bitwise_snapshot`; `unpruned_and_pruned_results_match_upstream_golden_fixtures` |
 | `remove_committee_forward_bonus` | `"let forward_bonus = if is_forward { 1.0 } else { 0.0 };"` → `"let forward_bonus = 0.0;"` | **EQUIVALENT.** The bonus is consulted only after all preceding rank components tie, and `FrontierCommittee::decode` already selects forward when `compare_committee_legs` returns `Equal`. Removing the bonus therefore leaves every binary-committee selection unchanged. |
 | `remove_active_mask_projection` | Delete `"and_assign(&mut state.active_syndrome, &column.active_mask);"` from `merge_branch`. | **EQUIVALENT.** A branch survives the close check only when every closing detector equals the observed bit. Keeping those now-inactive bits changes the noncanonical key representation but cannot prevent equivalent surviving prefixes from merging: all survivors carry the same observed values in those positions. |
 | `remove_transitions_increment` | Delete `"*transitions += 1;"` from `merge_branch`. | `transitions_count_every_candidate_branch_evaluation` |
 | `skip_dropped_mass_accumulation` | Delete `"dropped_log_mass = logaddexp(dropped_log_mass, scored.candidate.log_mass);"` from the discarded-candidate branch in `prune`. | `width_pruning_accounts_for_the_discarded_state_and_mass` |
-| `report_exact_after_pruning` | Replace the complete `"let status = if dropped_states == 0 { FrontierStatus::Exact } else { FrontierStatus::Pruned { k_capped, delta_pruned } };"` expression with `"let status = FrontierStatus::Exact;"`. | `width_pruning_accounts_for_the_discarded_state_and_mass`; `delta_pruning_reports_its_status_flag`; `one_prune_call_can_trigger_both_pruning_flags` |
+| `report_exact_after_pruning` | Replace the complete `"let status = if dropped_states == 0 { TrellisStatus::Exact } else { TrellisStatus::Pruned { k_capped, delta_pruned } };"` expression with `"let status = TrellisStatus::Exact;"`. | `width_pruning_accounts_for_the_discarded_state_and_mass`; `delta_pruning_reports_its_status_flag`; `one_prune_call_can_trigger_both_pruning_flags` |
 | `error_as_mismatch` | N/A: this mutation belongs to the `pecos-rslib` adapter layer, not `pecos-frontier`; this crate returns `DecoderError` directly and has no error-to-mismatch conversion. | N/A |
 | `merge_wrong_formula` | In `xor_combined_probability`, replace `"first * (1.0 - second) + second * (1.0 - first)"` with `"first * second"`. | `xor_probability_arithmetic_is_pinned`; `indistinguishable_merging_matches_unmerged_and_original_brute_force` |
 | `merge_keeps_last_occurrence` | Change `merge_indistinguishable_columns` so a repeated symptom tuple replaces the first retained tuple at the duplicate's later sequence position instead of updating the first tuple in place. | `merging_keeps_first_ordered_occurrence_and_deletes_later_copy` |
 | `merge_ignores_observables` | Change the `first_positions` key from the detector and observable word-mask pair to the detector word mask alone. | `merge_requires_matching_observable_sets` |
 | `bp_tables_from_dem_priors` | In `bp_suffix_compatibility`, replace the call that builds tables from the BP-derived `moments` with clones of each column's static `suffix_compatibility` table. | `bp_scores_change_the_greedy_survivor_without_changing_its_mass` |
 | `bp_clamp_removed` | In `bp_score_probability`, replace the complete `probability.clamp(BP_SCORE_PROBABILITY_MIN, 1.0 - BP_SCORE_PROBABILITY_MIN)` expression with `probability`. | `bp_score_probability_clamps_saturated_llrs` |
-| `bp_runs_in_unpruned_fast_path` | In `FrontierDecoder::from_sparse_dem`, delete `&& !(config.k == usize::MAX && config.delta.is_infinite())` from the `bp_score` construction condition. | `bp_flag_is_bitwise_inert_on_the_unpruned_fast_path` (including its `bp_seconds == 0.0` assertion) |
+| `bp_runs_in_unpruned_fast_path` | In `TrellisDecoder::from_sparse_dem`, delete `&& !(config.k == usize::MAX && config.delta.is_infinite())` from the `bp_score` construction condition. | `bp_flag_is_bitwise_inert_on_the_unpruned_fast_path` (including its `bp_seconds == 0.0` assertion) |
